Major update to board and GND pads
Added the Grove connectors.
Had to edit settings for the GND pads so that they tied to the copper pour with 'solid' instead of 'spokes' so that the pour areas would connect more. Also had to redefine the USB connector pins so that they were on the GND net. I'm getting some DRC errors that I think are erroneous, but I don't know why...